mostly changes forced by the junit stuff.
the testharness now creates a jar, smartfrog-testharness.jar , that contains junit base classes used by the component test routines, and so must be imported into the components before compiling their tests.

We really need to factor out this jar to not contain the tests, but that is a lower priority issue.